Split project with Blueprints
